create-devider-slider
