+++ introduction/home/templates/home/sections/introduction.html
@@ -17,6 +17,7 @@
                         height="{{ img.height }}" alt="{{ img.alt }}" class="w-100 h-100" />
                     </div>
                     <div class="col-9">
+                        <h1>{{ block.value.greeting }}</h1>
                         <p>{{ block.value.introduction_text|allow_safe_tags }}</p>
                     </div>
